added changes to take temparature, blood pressure readings in decimal

+++ Group3_ClinicDB/DAL/VisitsDAL.cs
@@ -326,12 +326,12 @@ namespace Group3_ClinicDB.DAL
             int appointment_id;
             decimal weight;
             decimal height;
-            int bodyTemparature;
+            decimal bodyTemparature;
             int pulse;
             string symptoms;
             int nurseId;
-            int bloodPressureSystolic;
-            int bloodPressureDiastolic;
+            decimal bloodPressureSystolic;
+            decimal bloodPressureDiastolic;
             string initialDiagnose;
             string finalDiagnose;
 
@@ -365,12 +365,12 @@ namespace Group3_ClinicDB.DAL
                             appointment_id = reader.GetInt32(appointmentIdOrd);
                             weight = reader.GetDecimal(weightOrd);
                             height = reader.GetDecimal(heightOrd);
-                            bodyTemparature = reader.GetInt32(bodyTemparatureOrd);
+                            bodyTemparature = reader.GetDecimal(bodyTemparatureOrd);
                             pulse = reader.GetInt32(pulseOrd);
                             symptoms = reader.GetString(symptomsOrd);
                             nurseId = reader.GetInt32(nurseIdOrd);
-                            bloodPressureSystolic = reader.GetInt32(bloodPressureSystolicOrd);
-                            bloodPressureDiastolic = reader.GetInt32(bloodPressureDiastolicOrd);
+                            bloodPressureSystolic = reader.GetDecimal(bloodPressureSystolicOrd);
+                            bloodPressureDiastolic = reader.GetDecimal(bloodPressureDiastolicOrd);
                             initialDiagnose = reader.GetString(initialDiagnoseOrd);
                             finalDiagnose = reader.GetString(finalDiagnoseOrd);
